+++ src/main/kotlin/core/command/PlayMusicCommand.kt
@@ -1,12 +1,10 @@
 package core.command
 
-import core.AudioScheduleResult
 import core.models.Event
 import core.models.MusicForPlayingParameters
 import core.models.Reply
 import core.music.BotAudioPlayer
 import music.TrackOrder
-import network.NetworkException
 
 /**
  * Команда воспроизведения музыки из URL или по ключевому слову.
@@ -20,7 +18,7 @@ class PlayMusicCommand(
 ) : Command(true) {
 
     override suspend fun invoke(event: Event): Reply {
-        val source = event.options.get("source") ?: return Reply.Text("Empty source.")
+        val source = event.options.get("source") ?: return Reply.Text("Источник не указан.")
         val result = botAudioPlayer.loadMusic(
             text = source,
             parameters = MusicForPlayingParameters(
@@ -28,19 +26,7 @@ class PlayMusicCommand(
                 guild = event.guild,
             )
         )
-        val text = when (result) {
-            is AudioScheduleResult.Error -> when (val ex = result.exception) {
-                is NetworkException -> "Ошибка при поиске на YouTube: ${ex.message}"
-                else -> "Error ${ex}."
-            }
-            AudioScheduleResult.NoMatches -> "No matches."
-            is AudioScheduleResult.PlaylistAdded -> "Playlist ${result.playlistName} loaded."
-            is AudioScheduleResult.PlaylistPlaying -> "Playlist ${result.playlistName} loaded. Now playing - ${result.trackName}."
-            is AudioScheduleResult.TrackAdded -> "Track ${result.trackName} added."
-            is AudioScheduleResult.TrackPlaying -> "Track ${result.trackName} playing."
-            is AudioScheduleResult.EmptyPlaylist -> "Playlist ${result.playlistName} is empty."
-            is AudioScheduleResult.TrackNotFound -> "Nothing was found for \"${result.keyword}\""
-        }
+        val text = result.toText()
         return Reply.Text(text)
     }
 }